Use lists instead of arrays in sequence repository usages
Lists perform better in performance, but worse in memory allocation than arrays in this case. For memory, it's not very significant (the power of 2 minus the item count is the approx. difference between them), so is the performance (approx. 25ns on 100 items), but for consistency and for low-cost optimization ToArray() were replaced with ToList().